requirements.txt
Keras==2.4.3
matplotlib==3.1.2
librosa==0.8.0
pandas==1.2.3
numpy==1.18.1
nltk==3.5
ipython==7.23.1
Pillow==8.2.0
scikit_learn==0.24.2
